473 add a mail cycle only destined to the quality control department

+++ src/Service/OperatorService.php
@@ -68,55 +68,55 @@ class OperatorService extends AbstractController
 
 
         // if ($today->format('d') % 4 == 0 && (!file_exists($filePath) || strpos(file_get_contents($filePath), $today->format('Y-m-d')) === false)) {
-        // if (!file_exists($filePath) || strpos(file_get_contents($filePath), $today->format('Y-m-d')) === false) {
+        if (!file_exists($filePath) || strpos(file_get_contents($filePath), $today->format('Y-m-d')) === false) {
 
-        $return = false;
+            $return = false;
 
-        $inActiveOperators = $this->operatorRepository->findOperatorWithNoRecentTraining();
-        // $this->logger->info('Inactive operators: ' . json_encode($inActiveOperators));
-        if (count($inActiveOperators) > 0) {
-            foreach ($inActiveOperators as $operator) {
-                $operator->setInactiveSince($today);
-                $this->em->persist($operator);
-            };
-            $this->em->flush();
-            $return = true;
+            $inActiveOperators = $this->operatorRepository->findOperatorWithNoRecentTraining();
+            // $this->logger->info('Inactive operators: ' . json_encode($inActiveOperators));
+            if (count($inActiveOperators) > 0) {
+                foreach ($inActiveOperators as $operator) {
+                    $operator->setInactiveSince($today);
+                    $this->em->persist($operator);
+                };
+                $this->em->flush();
+                $return = true;
+            }
+
+
+            $operatorSetToBeDeleted = $this->operatorRepository->findInActiveOperators();
+            // $this->logger->info('Inactive operatorSetToBeDeleted: ' . json_encode($operatorSetToBeDeleted));
+            if (count($operatorSetToBeDeleted) > 0) {
+                foreach ($operatorSetToBeDeleted as $operator) {
+                    $operator->setTobedeleted($today);
+                    $this->em->persist($operator);
+                };
+                $this->em->flush();
+                $return = true;
+            }
+
+
+            $toBeDeletedOperatorsIds = $this->operatorRepository->findOperatorToBeDeleted();
+            // $this->logger->info('Inactive toBeDeletedOperatorsIds: ' . json_encode($toBeDeletedOperatorsIds));
+            if (count($toBeDeletedOperatorsIds) > 0) {
+                foreach ($toBeDeletedOperatorsIds as $operatorId) {
+                    $this->entityDeletionService->deleteEntity('operator', $operatorId);
+                };
+                $this->em->flush();
+                $return = true;
+            }
+
+            if ($return) {
+                file_put_contents($filePath, $today->format('Y-m-d'));
+            }
+
+            $this->cache->delete('operators_list');
+
+            $countArray = [
+                'findDeactivatedOperators' => count($this->operatorRepository->findDeactivatedOperators()),
+                'toBeDeletedOperators' => count($toBeDeletedOperatorsIds)
+            ];
+            return $countArray;
         }
-
-
-        $operatorSetToBeDeleted = $this->operatorRepository->findInActiveOperators();
-        // $this->logger->info('Inactive operatorSetToBeDeleted: ' . json_encode($operatorSetToBeDeleted));
-        if (count($operatorSetToBeDeleted) > 0) {
-            foreach ($operatorSetToBeDeleted as $operator) {
-                $operator->setTobedeleted($today);
-                $this->em->persist($operator);
-            };
-            $this->em->flush();
-            $return = true;
-        }
-
-
-        $toBeDeletedOperatorsIds = $this->operatorRepository->findOperatorToBeDeleted();
-        // $this->logger->info('Inactive toBeDeletedOperatorsIds: ' . json_encode($toBeDeletedOperatorsIds));
-        if (count($toBeDeletedOperatorsIds) > 0) {
-            foreach ($toBeDeletedOperatorsIds as $operatorId) {
-                $this->entityDeletionService->deleteEntity('operator', $operatorId);
-            };
-            $this->em->flush();
-            $return = true;
-        }
-
-        if ($return) {
-            file_put_contents($filePath, $today->format('Y-m-d'));
-        }
-
-        $this->cache->delete('operators_list');
-
-        $countArray = [
-            'findDeactivatedOperators' => count($this->operatorRepository->findDeactivatedOperators()),
-            'toBeDeletedOperators' => count($toBeDeletedOperatorsIds)
-        ];
-        return $countArray;
     }
 }
-// }
